+++ job/steps/train_model.py
@@ -7,7 +7,7 @@ import pandas as pd
 import numpy as np
 
 from db.mappings.recommendation import Rec
-from job.steps.slow_knn import SlowKNN
+from job.steps.knn import KNN
 from job.steps.trainer import Trainer
 from lib.config import config
 
@@ -71,7 +71,7 @@ def get_recommendations(
     start_ts = time.time()
 
     # Use KNN similarity to calculate score of each recommendation
-    knn_index = SlowKNN(embeddings, df["date_decays"].values)
+    knn_index = KNN(embeddings, df["date_decays"].values)
     similarities, nearest_indices = knn_index.get_similar_indices(MAX_RECS + 1)
 
     knn_latency = time.time() - start_ts
